==> analysis/03_figures_and_tables.Rmd <==
---
title: "Figures & Tables"
site: workflowr::wflow_site
bibliography: ../data/references/references.bib
output:
  workflowr::wflow_html:
    toc: true
editor_options:
  chunk_output_type: console
---

```{r, setup, include=FALSE}
knitr::opts_chunk$set(
  echo = FALSE
)
```


# Introduction
This script generates the publication ready figures and supplementary tables

# Libraries and sources
These libraries and sources are used for this analysis.
```{r libs-and-src, message=FALSE, warning=FALSE, cache=FALSE, echo=TRUE}
# data wrangling
library(tidyverse)

# plotting
library(patchwork)
library(gtools)

# colors
library(AachenColorPalette)

# display tables
library(DT)

# development helpers
library(here)
library(devtools)

# other
library(openxlsx)

# source scripts
source(here("code/plots.R"))

# source scripts from external projects
base_url <- "https://raw.githubusercontent.com/saezlab/liver-disease-atlas"
source_url(file.path(base_url, "master", "code", "utils-plots.R"))
```

# Figures
```{r global-parameters}
fz <- 9
```

## Figure 1
### PCA
```{r pca-plot}
pca_result <- readRDS(here("output", "mouse-chronic-ccl4", "pca_result.rds"))

# limit pca ratio to one decimal digit
pca_result$var <- round(pca_result$var, 1)

pca_result$coords <- pca_result$coords |>
  mutate(treatment = str_to_title(treatment))

p1 <- plot_pca(pca_result, feature = "treatment") +
  scale_color_manual(
    values = aachen_color(c("turquoise", "orange")),
    labels = c(expression(paste(CCl[4])), expression(paste(Oil))),
    name = "Treatment"
  ) +
  theme(legend.text.align = 0) +
  my_theme(fsize = fz) +
  plot_annotation(
    title = "A",
    theme = theme(plot.title = element_text(size = fz + 2))
  )

p1
ggsave(here("figures/pca_plot.pdf"), p1, width = 3.5 * 1.2, height = 3.5)
```

### Top DEGs
```{r plot-top-degs, warning=FALSE}
degs <- readRDS(here("output", "mouse-chronic-ccl4", "genesets_degs.rds")) |>
  filter(contrast == "pure_ccl_12m_vs_0m") |>
  mutate(geneset_label = case_when(
    geneset == "ADH/ALDH" ~ "Alcohol and aldehyd dehydrogenases (ADH/ALDH)",
    geneset == "CYP" ~ "Cytochrome P450 dependent monooxygenases (CYP)",
    geneset == "DH" ~ "Dihydrodiol dehydrogenases (DH)",
    geneset == "EPH" ~ "Epoxid hydrolases (EPH)",
    geneset == "FMO" ~ "Flavin dependent monooxygenases (FMO)",
    geneset == "GST" ~ "Glutathione S-transferases (GST)",
    geneset == "MAO" ~ "Monoamine oxidases (MAO)",
    geneset == "MT" ~ "Methyltransferases (MT)",
    geneset == "NAT" ~ "N-Acetyltransferases (NAT)",
    geneset == "SULT" ~ "Sulfotransferases (SULT)",
    geneset == "UGT" ~ "UDP-glucuronosyltransferases (UGT)",
    geneset == "Blood" ~ "Blood side carrier",
    geneset == "Canalicular" ~ "Canalicular side carrier",
    TRUE ~ as.character(geneset)
  ))

deg_plots <- degs |>
  mutate(stars = stars.pval(fdr)) |>
  arrange(logFC) |>
  group_by(geneset, sign(logFC)) |>
  slice_max(abs(logFC), n = 10) |>
  add_count() |>
  ungroup() |>
  group_by(geneset) |>
  mutate(n = max(n)) |>
  ungroup() |>
  nest(data = -c(geneset, geneset_label, n, phase)) %>%
  mutate(p = pmap(., .f = plot_top_deg_genes, fontsize = 9))

deg_plots$p

plots <- pwalk(
  list(p = deg_plots$p, geneset = deg_plots$geneset, n = deg_plots$n), 
  .f = function(p, geneset, n, ...) {
    heights <- list(
      Blood = 2.05, Canalicular = 1.75, `ADH/ALDH` = 2.35, CYP = 3.1, DH = 1.55,
      EPH = 1.55, FMO = 1.75, GST = 2.55, MAO = 1.55, MT = 2.3, NAT = 2.3,
      SULT = 2.55, UGT = 2.35
    )

    if (geneset == "Blood") {
      p <- p + plot_annotation(
        title = "D   Carrier",
        theme = theme(plot.title = element_text(size = fz + 2))
      )
    } else if (geneset == "ADH/ALDH") {
      p <- p + plot_annotation(
        title = "B   Phase I Metabolism",
        theme = theme(plot.title = element_text(size = fz + 2))
      )
    } else if (geneset == "UGT") {
      p <- p + plot_annotation(
        title = "C   Phase II Metabolism",
        theme = theme(plot.title = element_text(size = fz + 2))
      )
    }

    # height = 1.05 + 0.2*n
    ggsave(
      filename = here(
        str_glue("figures/{str_replace(geneset, '/', '_')}.pdf")
      ),
      # filename = str_glue("~/Documents/figures/{str_replace(geneset, '/', '_')}.pdf"),
      plot = p,
      width = 3.5,
      height = heights[[as.character(geneset)]]
    )
  })
```

## Figure 2
### Single correlations
```{r single-correlation-plots}
corr <- readRDS(here("output", "mouse-chronic-ccl4", "correlations.rds")) |>
  filter(corr_fdr <= 0.1) |>
  arrange(p.value) |>
  unnest(data) |>
  mutate(anno = str_glue(
    "  r={round(estimate, 2)}\nFDR={signif(corr_fdr, digits=2)}"
  ))

p1 <- corr |>
  filter(key == "stage_6_vs_0 | nafld | hoang" & geneset == "DH") |>
  plot_correlation() +
  labs(
    x = "Human logFC (NAFLD Stage 6)",
    y = expression(paste("Mouse logFC (12 months ", CCl[4], ")", sep = "")),
    subtitle = "Dihydrodioldehydrogenases (DH)"
  ) +
  plot_annotation(
    title = "A",
    theme = theme(plot.title = element_text(size = fz + 2))
  )

p1
ggsave(here("figures/correlation_1.pdf"), p1, width = 3.5, height = 3.5)

p2 <- corr |>
  filter(key == "nash_vs_ctrl | omni | hampe14" & geneset == "NAT") |>
  plot_correlation() +
  labs(
    x = "Human logFC (NASH)",
    y = expression(paste("Mouse logFC (12 months ", CCl[4], ")", sep = "")),
    subtitle = "Acetyltransferases (NAT)"
  ) +
  plot_annotation(
    title = "B",
    theme = theme(plot.title = element_text(size = fz + 2))
  )

p2
ggsave(here("figures/correlation_2.pdf"), p2, width = 3.5, height = 3.5)

p3 <- corr |>
  filter(key == "nafld_vs_ctrl | omni | hampe14" & geneset == "ADH/ALDH") |>
  plot_correlation() +
  labs(
    x = "Human logFC (NAFLD)",
    y = expression(paste("Mouse logFC (12 months ", CCl[4], ")", sep = "")),
    subtitle = "Alcohol- and aldehyddehydrogenases (ADH/ALDH)"
  ) +
  plot_annotation(
    title = "C",
    theme = theme(plot.title = element_text(size = fz + 2))
  )

p3
ggsave(here("figures/correlation_3.pdf"), p3, width = 3.5, height = 3.5)
```

### Correlation heatmap
```{r correlation-heatmap}
annotation <- readRDS(
  here("data/meta-mouse-vs-human/contrast_annotation.rds")
) |>
  select(contrast, phenotype, source, label, disease)

corr <- readRDS(here("output", "mouse-chronic-ccl4", "correlations.rds")) |>
  separate(key, into = c("contrast", "phenotype", "source"), sep = " [|] ") |>
  left_join(annotation, by = c("contrast", "phenotype", "source")) |>
  complete(geneset, label, fill = list(estimate = NA_real_)) |>
  mutate(y = str_c(label, " | ", geneset)) |>
  mutate(x = str_c("CCl4 12 months| ", geneset)) |>
  mutate(highlight = case_when(
    geneset == "ADH/ALDH" & contrast == "nafld_vs_ctrl" & phenotype == "omni" &
      source == "hampe14" ~ "yes",
    geneset == "DH" & contrast == "stage_6_vs_0" & phenotype == "nafld" &
      source == "hoang" ~ "yes",
    geneset == "NAT" & contrast == "nash_vs_ctrl" & phenotype == "omni" &
      source == "hampe14" ~ "yes",
    TRUE ~ "no"
  )) |>
  mutate(stars = stars.pval(corr_fdr))

# heatmap of correlations
p4 <- corr |>
  ggplot(aes(x = geneset, y = label, fill = estimate)) +
  geom_tile(width = 0.95, height = 0.95) +
  geom_tile(
    data = filter(corr, highlight == "yes"), fill = NA,
    color = aachen_color("red"), size = 0.5
  ) +
  scale_fill_gradient2(
    low = aachen_color("orange"),
    mid = "white",
    high = aachen_color("green"),
    na.value = aachen_color("black25"),
    breaks = c(-1, 0, 1), labels = c(-1, 0, 1),
    limits = c(-1, 1)
  ) +
  scale_color_manual(values = c("white", aachen_color("red"))) +
  geom_text(
    data = filter(corr, stars != "."), aes(label = stars),
    vjust = 1, size = fz / (14 / 5)
  ) +
  geom_text(
    data = filter(corr, stars == "."), aes(label = stars),
    vjust = 0, size = fz / (14 / 5)
  ) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(x = "Mouse", y = "Human") +
  guides(fill = guide_colorbar("r"), color = "none") +
  theme(axis.line = element_blank(), axis.ticks = element_blank()) +
  my_theme(grid = "no", fsize = fz) +
  plot_annotation(
    title = "D",
    theme = theme(plot.title = element_text(size = fz + 2))
  )

p4
ggsave(here("figures/heatmap.pdf"), p4, width = 5.3, height = 4)
```

### Top DEGs
```{r plot-top-degs-hs, warning=FALSE}
gene_mapping <- readRDS(here("data/annotation/gene_id_annotation.rds")) |>
  distinct(gene = symbol_mgi, symbol_hgnc) |>
  # remove predicted genes that start with Gm followed by digits
  filter(!str_detect(gene, "^Gm\\d+")) |>
  drop_na()

degs_hs <- readRDS(here("output", "mouse-chronic-ccl4", "genesets_degs_hs.rds")) |>
  filter(contrast == "stage_6_vs_0") |>
  mutate(geneset_label = case_when(
    geneset == "ADH/ALDH" ~ "Alcohol and aldehyd dehydrogenases (ADH/ALDH)",
    geneset == "CYP" ~ "Cytochrome P450 dependent monooxygenases (CYP)",
    geneset == "DH" ~ "Dihydrodiol dehydrogenases (DH)",
    geneset == "EPH" ~ "Epoxid hydrolases (EPH)",
    geneset == "FMO" ~ "Flavin dependent monooxygenases (FMO)",
    geneset == "GST" ~ "Glutathione S-transferases (GST)",
    geneset == "MAO" ~ "Monoamine oxidases (MAO)",
    geneset == "MT" ~ "Methyltransferases (MT)",
    geneset == "NAT" ~ "N-Acetyltransferases (NAT)",
    geneset == "SULT" ~ "Sulfotransferases (SULT)",
    geneset == "UGT" ~ "UDP-glucuronosyltransferases (UGT)",
    geneset == "Blood" ~ "Blood side carrier",
    geneset == "Canalicular" ~ "Canalicular side carrier",
    TRUE ~ as.character(geneset)
  )) |>
  # map back to human symbols
  inner_join(gene_mapping, by = "gene") |>
  # select(-gene) |>
  # rename(gene = symbol_hgnc)
  mutate(gene = case_when(
    str_to_upper(gene) == symbol_hgnc ~ symbol_hgnc,
    TRUE ~ paste0(symbol_hgnc, " (", gene, ")")
    # str_to_upper(gene) != symbol_hgnc ~ str_c(symbol_hgnc, "(", gene, ")", collapse = "")
  ))

deg_plots_hs <- degs_hs |>
  mutate(stars = gtools::stars.pval(fdr)) |>
  arrange(logFC) |>
  group_by(geneset, sign(logFC)) |>
  slice_max(abs(logFC), n = 10) |>
  add_count() |>
  ungroup() |>
  group_by(geneset) |>
  mutate(n = max(n)) |>
  ungroup() |>
  nest(data = -c(geneset, geneset_label, n)) %>%
  mutate(p = pmap(., .f = plot_top_deg_genes, fontsize = 9))

deg_plots_hs$p

plots_hs <- pwalk(
  list(p = deg_plots_hs$p, geneset = deg_plots_hs$geneset, n = deg_plots_hs$n),
  .f = function(p, geneset, n, ...) {
    heights <- list(
      Blood = 1.55, Canalicular = 1.85, `ADH/ALDH` = 2.55, CYP = 3.6, DH = 1.75,
      EPH = 1.55, FMO = 1.59, GST = 1.59, MAO = 1.55, MT = 2.4, NAT = 1.55,
      SULT = 1.55, UGT = 1.55
    )

    if (geneset == "Canalicular") {
      p <- p + plot_annotation(
        title = "F   Carrier",
        theme = theme(plot.title = element_text(size = fz + 2))
      )
    } else if (geneset == "CYP") {
      p <- p +
        plot_annotation(
          title = "E   Phase I Metabolism",
          theme = theme(plot.title = element_text(size = fz + 2))
        ) & theme(axis.text = element_text(size = fz - 1))
    } else if (geneset == "ADH/ALDH") {
      p <- p + plot_annotation(
        title = "A   Phase I Metabolism",
        theme = theme(plot.title = element_text(size = fz + 2))
      )
    } else if (geneset == "MT") {
      p <- p + plot_annotation(
        title = "B   Phase II Metabolism",
        theme = theme(plot.title = element_text(size = fz + 2))
      )
    }

    # height = 1.05 + 0.2*n
    ggsave(
      filename = here(
        str_glue("figures/{str_replace(geneset, '/', '_')}_hs.pdf")
      ),
      plot = p,
      width = 3.5,
      height = heights[[as.character(geneset)]]
    )
  })
```

# Tables
## List of metabolic and carrier genes with degs
```{r metabolic-carrier-genes}
degs <- readRDS(here("data", "mouse-chronic-ccl4", "limma_result.rds")) |>
  filter(contrast == "pure_ccl_12m_vs_0m")

genes <- readRDS(here("output/mouse-chronic-ccl4/metabolic_genes.rds")) |>
  left_join(degs) |> 
  select(gene, family, phase, logFC, pval, fdr) 
carrier <- readRDS(here("output/mouse-chronic-ccl4/carrier_genes.rds")) |>
  left_join(degs) |> 
  select(gene, side, logFC, pval, fdr)

datatable(genes)
datatable(carrier)

tibble(sheet_name = c("Metabolic genes", "Carrier genes"), data = list(genes, carrier)) |>
  deframe() |>
  write.xlsx(here("tables/metabolic_carrier_genes.xlsx"),
    keepNA = TRUE
  )
```
